+++ test/musubi_test.exs
@@ -1,3 +1,5 @@
 defmodule MusubiTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
+
+  doctest Musubi
 end